backend/github_service/requirements.txt

PyGithub==2.1.1
python-dotenv==1.0.0
requests==2.31.0
